Create UI component for presets and improve preset manager (#54)
* Add PresetsManager to PluginBase

* Start setting up preset component

* Add escape option to preset name editor

* Add next/prev arrows to preset menu

* Add API for loading user presets with backwards compatibility

* Apply clang-format

* Fix compilation warnings/errors

* Apply clang-format

* Update tests for preset behaviour

* Add PresetsComp to GUITest

* Add more unit tests for presets menu

* Apply clang-format

* Fix unit test compilation

* Add more tests for presets component

* Apply clang-format

* Skip presets comp tests that need filechooser on Linux

* Re-add other unit tests

* Move presetMap from std::unordered_map to std::map

* Minor-ish tweaks to presets behaviour

* Fix unit tests

+++ PluginUtils/Presets/chowdsp_PresetManager.cpp
@@ -80,6 +80,9 @@ const Preset* PresetManager::findPreset (const Preset& presetToFind) const
 
 void PresetManager::setDefaultPreset (Preset&& newDefaultPreset)
 {
+    // default preset must be a valid preset!
+    jassert (newDefaultPreset.isValid());
+
     auto* foundDefaultPreset = findPreset (newDefaultPreset);
 
     if (foundDefaultPreset != nullptr)
@@ -107,17 +110,14 @@ void PresetManager::saveUserPreset (const juce::File& file)
     auto stateXml = savePresetState();
     const auto name = file.getFileNameWithoutExtension();
 
-    int presetID = userIDMap["User"];
-    while (presetMap.find (presetID) != presetMap.end())
-        presetID++;
+    keepAlivePreset = std::make_unique<Preset> (name, "User", *stateXml.get());
+    if (keepAlivePreset != nullptr)
+    {
+        keepAlivePreset->toFile (file);
+        loadPreset (*keepAlivePreset);
 
-    auto newPreset = Preset (name, "User", *stateXml.get());
-    auto [iterator, success] = presetMap.insert ({ presetID, std::move (newPreset) });
-
-    jassert (success);
-    auto& preset = iterator->second;
-    preset.toFile (file);
-    loadPreset (preset);
+        loadUserPresetsFromFolder (getUserPresetPath());
+    }
 }
 
 void PresetManager::setIsDirty (bool shouldBeDirty)
@@ -143,13 +143,23 @@ void PresetManager::loadPreset (const Preset& preset)
 
 std::unique_ptr<juce::XmlElement> PresetManager::savePresetState()
 {
-    return vts.state.createXml();
+    auto xml = vts.state.createXml();
+    xml->deleteAllChildElementsWithTagName (presetStateTag);
+
+    JUCE_BEGIN_IGNORE_WARNINGS_GCC_LIKE ("-Wpessimizing-move")
+    return std::move (xml);
+    JUCE_END_IGNORE_WARNINGS_GCC_LIKE
 }
 
 void PresetManager::loadPresetState (const juce::XmlElement* xml)
 {
     auto newState = juce::ValueTree::fromXml (*xml);
     vts.replaceState (newState);
+}
+
+Preset PresetManager::loadUserPresetFromFile (const juce::File& file)
+{
+    return { file };
 }
 
 int PresetManager::getIndexForPreset (const Preset& preset) const
@@ -197,6 +207,9 @@ juce::File PresetManager::getUserPresetConfigFile() const
 
 void PresetManager::setUserPresetPath (const juce::File& file)
 {
+    if (file == juce::File())
+        return;
+
     auto config = getUserPresetConfigFile();
     config.deleteFile();
     config.create();
@@ -216,8 +229,8 @@ juce::File PresetManager::getUserPresetPath() const
 void PresetManager::loadUserPresetsFromFolder (const juce::File& file)
 {
     std::vector<Preset> presets;
-    for (auto& f : file.findChildFiles (juce::File::findFiles, true))
-        presets.emplace_back (f);
+    for (const auto& f : file.findChildFiles (juce::File::findFiles, true))
+        presets.push_back (loadUserPresetFromFile (f));
 
     // delete old user presets
     int presetID = userIDMap["User"];
@@ -256,9 +269,9 @@ void PresetManager::loadXmlState (juce::XmlElement* xml)
         return;
     }
 
-    statePreset = std::make_unique<Preset> (xml->getChildByName (Preset::presetTag));
-    if (statePreset != nullptr)
-        loadPreset (*statePreset);
+    keepAlivePreset = std::make_unique<Preset> (xml->getChildByName (Preset::presetTag));
+    if (keepAlivePreset != nullptr)
+        loadPreset (*keepAlivePreset);
 
     setIsDirty ((bool) xml->getIntAttribute (presetDirtyTag));
 }
